build(deployment): local deployment
Motivation
----------
On the weekend we had a downtime on production because `nginx` configurations got out of sync with the code in newer releases.

To prevent this and further downtimes I suggest to check in our deployment configuration and make it *reproducible*.

This PR is an attempt to create a docker container that behaves as similar as possible to production and can be used as a sandbox for changes to our deployment configuration.

Read about infrastructure as code: https://en.wikipedia.org/wiki/Infrastructure_as_code

How to test
-----------
1. `cd deployment`
2. `docker build . --tag dreammall-deployment`
3. `docker run -it -p 8080:80 dreammall-deployment sh` .. and wait
4. Following services are running:
  * http://localhost:8080/
  * http://localhost:8080/api
  * http://localhost:8080/docs
5. Be amazed