feat(authentik): non-binary dump
Motivation
----------
See #888 - we want to be able to see the plaintext seeds and we want to
see the delta between versions when the seeds change.

This will implement seed data as SQL dumps. I created the dump by
executing the `unpack` script on `master` and `pg_dump` from within the
docker container.

The `README.md` now contains instructions how to update the seeds and
how to restore the seeds dump.

Unfortunately it is not possible to only dump certain tables with
`pg_dump --table <table_name>` because there are foreign key constraints
which `pg_dump` is not able to resolve.

Side quest of this PR:
* `docker compose down -v` to wipes out the authentik database for
  a clean start
* no more permission denied errors because the docker user (running as
  `root`) created the `authentik/database/` bound volume.

How to test
-----------
1. Follow the README

close #888