refactor(docker): remove unused volumes (#1365)
Motivation
----------
These volumes are not in use, let's remove them. As the comment says,
the original intention was to have a different node version in the
container and the host.

This is something that we *don't* want. We want to use the same node
version *always*, see #1319.

How to test
-----------
1. Nothing to test, disabled code.