build(workflow): same node version on all stages (#1319)
Motivation
----------
We have completely different `node --version`s on local development,
build server (even in different jobs) and production.

My suggestion is that we configure the node version in a central file in
the root folder and reference it from everywhere.

For me, any version manager is fine. If I have the choice, I suggest
[asdf-vm]](https://github.com/asdf-vm/asdf).

How to test
-----------
1. Check status checks