feat(workflows): auto-assign PR author
Motivation
----------
I've heard today that we want to assign each PR to its respective
author, so the name appears in the PR list on the right side:
https://github.com/dreammall-earth/dreammall.earth/pulls

I keep forgetting this, so I'd say let's drop the requirement or
automate this.

Feel free to reject this PR in case we drop the requirement.

How to test
-----------
1. Merge this into `master`
2. See if new PRs get auto-assigned