fix(other): nginx configuration for websockets
Motivation
----------
The websocket connection to the backend introduced in #1268 does not
work because of a unluckly nginx configuration on the vServer in
production/staging.

In #1402 I could reproduce the problem locally. After numerous
tries/errors with the nginx configuration I'm now pretty sure this will
enable websockets on production.

How to test
-----------
1. Ensure that the `frontend/.env` has the exact same values as the
   currently checked in `frontend/.env.dist`:
   ```
    # The trailing slash is important!
    PUBLIC_ENV__ENDPOINTS__GRAPHQL_URI=http://localhost:4000/
    # The *missing* trailing slash is important!
    PUBLIC_ENV__ENDPOINTS__WEBSOCKET_URI=ws://localhost:4000/subscriptions
   ```
2. (optional) If you had to change `.env` values, run a `deployment/deploy.sh`
3. Checkout this feature branch, copy `deployment/nginx/default.conf`
   into `/etc/nginx/http./default.conf`
4. `rc-service nginx restart`
5. Works :)