fix(backend): endless redirects
Motivation
----------
This is a side quest of #1451.

If there is an error with the prisma client, we send `UNAUTHENTICATED`
instead of `INTERNAL_ERROR`. This will lead to an endless loop because
the frontend redirects to `/signin` over and over again.

I also made a refactoring: I want to be able to mock the prisma client,
so that it rejects with an error. It is a best practice to use the
`dataSources` pattern from Apollo <= v2:
https://www.apollographql.com/docs/apollo-server/v2/data/data-sources/

In Apollo v4 you simply put it into `context`. I like the name
`dataSources` though and nested the `prisma` client in there.

How to test
-----------
1. `docker compose down database -v`
2. `docker compose up database backend frontend authentik`
3. Login in at http://localhost:3000/
4. No endless redirects

ToDos
-----
As a follow up we should make sure to always use the `prisma` client
from the `context`. But I'll leave this for another day.