refactor(docker): reduce production image size (#1501)
Motivation
----------
Side quest of #1451. This will reduce the docker images sized for
`--target production`. Typescript should be a development dependency for
sure and `ts-node` is not used as the default `CMD` in the `Dockerfile`
but `node ./build/index.cjs`.

On our current production instance we install development dependencies
as well, so this shouldn't break anything there.

The `production` docker images are not used in a deployment as of now, but are
relevant for #1451.

How to test
-----------
In each subfolder:
1. `npm install --omit=dev`
2. `du -sh node_modules`
3. E.g. in `frontend/` it went from `726M` down to `194M`
3. In `presenter/` it went from `688M` down to `179M`

Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>