Modify LogoImage test to see if it fixes race condition bug (#1677)
We still have the VImg test error sometimes, which is annoying and slows down development because CI blocks and you are never sure if you just introduced a new error.

This is an attempt to fix it. I examined all test files where VImg is used without an VApp context, and this file is the only one I found.

If it doesn't fix the problem, it won't hurt either.