refactor(frontend): mock apollo in each test suite (#1740)
Motivation
----------
This is a side quest of #1689:

@Bettelstab and I figured out in a pairing that our frontend tests are
not isolated. State is mutated in one test and affects the behaviour of
a completely unrelated test.

This is a first step: If a test suite uses apollo client, the test suite
is responsible to mock all the data. No global state anymore.

How to test
-----------
1. The behaviour should not have changed.