fix(authentik): email enrollment
Motivation
----------
Before we knew authentik blueprints we maintained two flows: One without
email delivery (for local development probably) and one in production
with email delivery.

But we use Mailpit now. So we can just merge the two flows.

Second, the `dreammallearth-enrollment-email` used to have SMTP
configurations stored in the blueprint. That's a big no-no. The SMTP
configurations are handled via environment variables passed to the authentik
worker pod/container.

How to test
-----------
1. `docker compose down -v`
2. Visit http://localhost:9000/if/admin/#/flow/flows/dreammallearth-enrollment-email
3. Click on "Execute Flow >> Normal"
4. Register
5. Find E-Mail in Mailpit http://localhost:8025/
6. Click signup link in E-Mail
7. You are logged in! 🥳