feat(frontend): complete a subscription
Motivation
----------
Who: Users
What: Click on a button "Complete a subscription" to send money on a monthly basis to the company running "DreamMall".
Why: Because I want to support the development of this wonderful piece of software as well as the hosting and administration of it.

-----------
1. Visit http://localhost:3000/cockpit
2. Click on "Upgrade to premium"
3. See instructions how to make a bank transfer

You can change the amount and it will update the QR code.

There is another refactoring which affects the cockpit as a whole:
I removed the `narrow` property which is unnecessary.

+ Increased max width and height for the cockpit cards

close #1030