feat(backend): unique reference for bank transfer (#1936)
* feat(backend): unique reference for bank transfer

Motivation
----------
Who: Admins
What: Give each user account unique and immutable ID or reference that can be used as a "purpose" on bank transfers.
Why: To be able to assign a payment to a user account.

How to test
-----------
1. `cd backend`
2. `npm run test:unit`

close #1909

* fix: handle existing data

Prisma's lack of support for data migrations is a joke:
https://www.prisma.io/docs/orm/prisma-migrate/workflows/data-migration

* fix(backend): drop reference_id default

I see strange behaviour: Every time I want to run `npm run db:migrate`
prisma always wants to create a new migration. I fixed this by removing
the default as soon as the first migration has been applied. This is
even a good thing! But I find the behaviour of Prisma very weird if you
use `@default(dbgenerated())`.