feat(backend): check user right to get meeting information.
Motivation
----
If a user set a meeting to private he want to be sure that no other user as the ones he has invited to his meeting.
Filters meetings that are public or user.id is in meeting.users.

Test
----
Create a private table invite multiple users open the table, on another tab delete the table from the my tables section.
Now this table is not present in the table side bar.