refactor(backend): unrelated refactoring
Motivation
----------
This is a side-quest for #2141. I thought I would need this and respond an in-memory welcome `Table` or `OpenTable` in the table resolver.

I changed my mind and would like to implement as much as possible in the frontend itself without adding complexity to the backend.

This PR contains a refactoring that I thought was necessary but isn't. But it doesn't hurt either. Feel free to approve or reject this PR.

How to test
-----------
1. Check build server status