refactor(backend): inject prisma via contextValue (#2329)
Motivation
----------
We already have `prisma` as `dataSource` in our `ContextValue`. This PR
will import them in our resolvers.

We're not fully done yet, there are imports of `prisma` elsewhere.
However, it's good for now. The code of the resolvers get copy+pasted,
so it's important to refactor this part already.

How to test
-----------
1. `npm run test:unit`
2. All tests pass