refactor(backend): pass CONFIG in contextValue (#2327)
Motivation
----------
This is the remaining part of #2287.

If we pass `CONFIG` via dependency injection in `contextValue` we don't have to mock entire javascript modules in the tests. Instead, we can set dependencies per test case.

For `BBB` it would make sense to mock `axios` for example instead of our own code. Always try to mock as little as possible.

How to test
-----------
1. `npm run test:unit`