feat(backend): change openTables query to give back three lists
The design needs a split of the tables (Welcome, MallTalk & Project)
The table needs to know which table type it is and if the user is moderator for the meeting.

Change the openTables query so that it gives back an OpenTables object.
Change the joinTable mutation so that it gives back an JoinTable object.