refactor(frontend): remove global mocks in tests (#2444)
Motivation
----------
This removes the global mock of `oidc-client-ts`.

Mocking globally will affect *every* test and all of a sudden importing the module will behave differently. It is not explicitly done in the test that depend on the behaviour, adding confusion and potential for bugs.

This complexity is now refactored with good-old dependency injection. No need for mocks.

Furthermore, this tests collects every global mutation of `@vue/test-utils` in a single file to better communicate what to expect when you create a component wrapper with this framework.

How to test
-----------
1. `cd frontend && npm run test:unit -- --no-coverage`
2. Same same as before. But less code!