refactor(other): eslint no-catch-all
Motivation
----------
This PR will prevent errors from silently failing. It's a common bug to have a `try/catch` and some error thrown which is *not* the one the developer was looking for.

Have a look at this PR for example: https://github.com/dreammall-earth/dreammall.earth/pull/2405#pullrequestreview-2333456869

close #1548

How to test
-----------
1. Write a `try/catch` where you don't rethrow the `error`
2. `npm run test:lint:eslint`
3. Will fail