feat(backend): log unexpected brevo api responses
Motivation
----------
So, this should fix: #2680

I was not able to reproduce the error in a manual E2E test. @ulfgebhardt do you know how to reproduce the catch block with Brevo?

I hoped to "unsubscribe" to a newsletter and resubscribe again to get into the catch block, but that would only trigger the "happy path" (status code 2XX).

How to test
-----------
1. `cd backend`
2. `npm run test:unit`
3. Tests pass