feat(backend): database add invitation links related to a owner and a redeemer
Motivation
---
We need to store the invitation links a user want to send out and relate it with the user that redeemed it.

Solution
---
Create an InvitationLink table, to store the created link, created code and acceptedUserId must be unique for acceptedUser and the Id of it can be null.
Add to the User a list of owned InvitationLink and a possible redeemed link.

Issues
---
- relates #2255