feat(backend): createInvitationLink mutation
Motivation
----
The frontend has a button for the user to get an invitation link so we need to create a link and store it in the database.

Solution
----
Added a createInvitationLink mutation that creates an unique invitationCode, and stores it.
You can call the mutation like the following:
```
mutation {
  createInvitationLink
}
```
And will get an object like this:
```
{
  "data": {
    "createInvitationLink": "http://localhost:3000/invite/fd887080-2176-4c02-b457-7de98e83e36f"
  }
}
```

Issues
----
- relates #2255